Moola adapter - cleanup -
